requirements.txt
Django==2.2.5
djangorestframework==3.13.1
redis==4.1.4